docs/legendwon/oauth2.0_code-main/requirements.txt
Flask==3.0.0
Authlib==1.3.0
Werkzeug==3.0.1
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
cryptography==41.0.7
